pyproject.toml
[tool.poetry]
name = "sfdata-schema"
version = "0.1.0"
description = "A basic structure for documenting tabular data"
authors = ["Kaj Siebert <kaj@k-si.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyhumps = "^3.8.0"
PyYAML = "^6.0"
jstyleson = "^0.0.2"
Jinja2 = {version = "^3.1.2", optional = true}
graphviz = {version = "^0.20.1", optional = true}

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
coverage = "^6.5.0"
pytest = "^7.1.3"

[tool.poetry.extras]
erd = ["Jinja2", "graphviz"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
